requirements.txt
annotated-types==0.7.0
asgiref==3.8.1
Django==5.1.2
django-bootstrap-datepicker-plus==5.0.5
django-environ==0.11.2
psycopg2==2.9.10
pydantic==2.9.2
pydantic_core==2.23.4
sqlparse==0.5.1
typing_extensions==4.12.2
